Renaming queues, renami\ng 'gdpr' to 'datautility', renaming 'function' to 'functions'
